+++ CHANGES.md
@@ -6,6 +6,11 @@
 * The form associated elements returned by `FormElement.elements()` now reflect changes made to the DOM, 
   subsequently to the original parse. [2140](https://github.com/jhy/jsoup/issues/2140)
 
+### Bug Fixes
+
+* `Element.cssSelector()` would fail if the element's class contained a `*`
+  character. [2169](https://github.com/jhy/jsoup/issues/2169)
+
 ## 1.18.1 (2024-Jul-10)
 
 ### Improvements